requirements.txt
fastapi[all]
motor
requests
pytest
passlib
bcrypt
pyjwt
python-jose